Moving DalitzPlotPdf to Amp3Body
