+++ include/goofit/PDFs/physics/AmpNBody.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <goofit/PDFs/GooPdf.h>
+
+namespace GooFit {
+
+class AmpNBody : public GooPdf {
+public:
+    using GooPdf::GooPdf;
+};
+
+
+} // namespace GooFit
