+++ python/PDFs/physics/AmpNBody.cpp
@@ -1,0 +1,11 @@
+#include <goofit/Python.h>
+
+#include <goofit/PDFs/physics/AmpNBody.h>
+
+using namespace GooFit;
+
+void init_AmpNBody(py::module &m) {
+    py::class_<AmpNBody, GooPdf>(m, "AmpNBody")
+
+        ;
+}
